third_party/blink/web_tests/external/wpt/css/cssom/getComputedStyle-detached-subtree-expected.txt

This is a testharness.js-based test.
PASS getComputedStyle returns no style for detached element
FAIL getComputedStyle returns no style for element in non-rendered iframe (display: none) assert_equals: expected 0 but got 335
FAIL getComputedStyle returns no style for element in non-rendered iframe (display: none) from iframe's window assert_equals: expected 0 but got 335
FAIL getComputedStyle returns no style for element outside the flat tree assert_equals: expected 0 but got 335
FAIL getComputedStyle returns no style for descendant outside the flat tree assert_equals: expected 0 but got 335
PASS getComputedStyle returns no style for shadow tree outside of flattened tree
Harness: the test ran to completion.
